rag_pinecone_project/requirements.txt
streamlit
sentence-transformers
transformers
torch
pinecone
PyPDF2
python-docx
